backend/requirements.txt
alembic==1.8.1
anyio==3.6.2
attrs==22.1.0
autopep8==2.0.0
bcrypt==4.0.1
black==22.10.0
certifi==2022.9.24
cffi==1.15.1
click==8.1.3
cryptography==38.0.3
ecdsa==0.18.0
fastapi==0.87.0
h11==0.14.0
httpcore==0.16.1
httptools==0.5.0
httpx==0.23.1
idna==3.4
iniconfig==1.1.1
Mako==1.2.4
MarkupSafe==2.1.1
mypy-extensions==0.4.3
numpy==1.23.5
packaging==21.3
pandas==1.5.1
passlib==1.7.4
pathspec==0.10.2
platformdirs==2.5.4
pluggy==1.0.0
psycopg2==2.9.5
pyasn1==0.4.8
pycodestyle==2.9.1
pycparser==2.21
pydantic==1.10.2
pyparsing==3.0.9
pytest==7.2.0
python-dateutil==2.8.2
python-dotenv==0.21.0
python-jose==3.3.0
python-multipart==0.0.5
pytz==2022.6
PyYAML==6.0
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.44
starlette==0.21.0
tomli==2.0.1
typing_extensions==4.4.0
uvicorn==0.19.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
